requirements.txt
requests
pyyaml
json
zeep